:tractor: Reorganize structure
